sunRadius
The radius of the sun and moon circles is now scaled to the width of the image. These values were previously constant and appropriate for the default width, however devices with larger dpi would upscale the width to consume available screenspace (producing circles that were disproportionately small).

sunRadius refactored to sunScale; moonRadius refactored to moonScale. issue #286