repeatingDays
The repeatingDays option uses the eventTime (and not the adjusted alarmTime).

E.g. The alarm is set to 8h before sunrise.. suppose today is Sunday and the alarm is set to repeat on mon, tue, wed, thu, fri. The event time is sunrise 7:30 (mon), but the adjusted alarmTime is 11PM (previous day). In this example an alarm will sound Sunday at 11PM (even though sunday isn't one of the repeating days) because the repeating days uses the event time (monday) and not the adjusted time (sunday).