hardware buttons
Hides the hardwarebutton pref and changes the default action from silent to snooze.

The button presses are only detected from the fullscreen activity, and not caught from the lockscreen or when the screen is off. Additional work is required to get these keypresses. The pref is hidden until then to avoid advertising a feature that appears broken.