refactor
Refactors methods and enums; "vernal" to "spring". The phrase "vernal equinox" is inaccurate for these methods, which are adjusted to the location's hemisphere by the calculator.  The vernal equinox has a very specific meaning - it is the spring equinox in the northern hemisphere, first point of aries, start of sidereal year.

Increments calculator version; 1.5.0 -> 1.5.1
Deprecates `getVernalEquinoxForYear`, replaced with `getSpringEquinoxForYear`