test
Adds failing tests for CalculatorProvider (exposes a time zone related bug in the `queryMoon` method).

The test fails because the provider fails to apply the user selected timezone (mismatch between actual and expected results). On the chosen test day (2020-04-11) moonrise occurs on the 11th but not the 12th for US/Arizona (user selected), and vice versa for MST (+1hr dst; system).