updateAlarmTime
The `updateAlarmTime` method should not modify `item.hour` or `item.minute` when scheduling "clock time" alarms. These alarms are a special case where the hour and minute are user-defined (and should remain unchanged).

For most `TimeZone` objs the modification is harmless (replaced with same values), but when alarms are configured to `ApparentSolarTime` this causes the user-defined time to drift every time the alarm is (re)scheduled (#715).