location permission warning
Show a warning message when the app is configured to "current location" mode but location permissions are denied. #733

Permission is automatically requested when a location update occurs, but if the user repeatedly denies permission then future requests fail silently (and "current location" mode seems broken). This warning message is a fallback to alert users that something is wrong with the app configuration.